+++ Configuration GUI/src/common/Definitions.ts
@@ -29,12 +29,11 @@ export const auto_neg_mapping: { [name: string]: string } = {
 }
 
 export const speed_mapping: { [name: string]: string } = {
-    "BF_SPEED_1G": "1G",
     "BF_SPEED_10G": "10G",
     "BF_SPEED_25G": "25G",
     "BF_SPEED_40G": "40G",
-    "BF_SPEED_50G": "50G",
-    "BF_SPEED_100G": "100G"
+    "BF_SPEED_100G": "100G",
+    "BF_SPEED_400G": "400G"
 }
 
 export const loopback_mapping: { [name: string]: string } = {