+++ Configuration GUI/src/sites/Ports.tsx
@@ -23,14 +23,9 @@ import { get, post } from "../common/API";
 import { Button, Col, Form, Row, Table } from "react-bootstrap";
 import styled from "styled-components";
 import InfoBox from "../components/InfoBox";
-import { P4TGConfig } from "../common/Interfaces";
-import {
-  auto_neg_mapping,
-  fec_mapping,
-  loopback_mapping,
-  speed_mapping,
-} from "../common/Definitions";
-import { GitHub } from "./Home";
+import {ASIC, FEC, P4TGConfig, P4TGInfos, SPEED} from "../common/Interfaces";
+import {auto_neg_mapping, fec_mapping, loopback_mapping, speed_mapping} from "../common/Definitions";
+import {GitHub} from "./Home";
 import translate from "../components/translation/Translate";
 
 const StyledCol = styled.td`
@@ -56,10 +51,10 @@ export const PortStatus = ({ active }: { active: boolean }) => {
   );
 };
 
-const Ports = () => {
-  const [loaded, set_loaded] = useState(false);
-  const [ports, set_ports] = useState([]);
-  const [config, set_config] = useState<P4TGConfig>({ tg_ports: [] });
+const Ports = ({p4tg_infos}: {p4tg_infos: P4TGInfos}) => {
+    const [loaded, set_loaded] = useState(false)
+    const [ports, set_ports] = useState([])
+    const [config, set_config] = useState<P4TGConfig>({tg_ports: []})
 
   const loadPorts = async () => {
     let stats = await get({ route: "/ports" });
@@ -89,57 +84,57 @@ const Ports = () => {
     });
 
     if (update.status === 201) {
-      refresh();
-    }
-  };
+      refresh()
+  }
+}
 
-  const updateArp = async (pid: number, state: boolean) => {
-    console.log(state);
-    let update = await post({
-      route: "/ports/arp",
-      body: {
-        pid: pid,
-        arp_reply: state,
-      },
-    });
-
-    if (update.status === 201) {
-      refresh();
-    }
-  };
-
-  const getMac = (port: number) => {
-    let mac = "Unknown";
-
-    config.tg_ports.forEach((p) => {
-      if (p.port == port) {
-        mac = p.mac;
+const updateArp = async (pid: number, state: boolean) => {
+  console.log(state)
+  let update = await post({
+      route: "/ports/arp", body: {
+          pid: pid,
+          arp_reply: state
       }
-    });
+  })
 
-    return mac;
-  };
+  if (update.status === 201) {
+      refresh()
+  }
+}
 
-  const getArpReply = (port: number) => {
-    let reply = false;
+const getMac = (port: number) => {
+  let mac = "Unknown"
 
-    config.tg_ports.forEach((p) => {
-      if (p.port == port) {
-        reply = p.arp_reply ?? false;
+  config.tg_ports.forEach(p => {
+      if(p.port == port) {
+          mac = p.mac
       }
-    });
+  })
 
-    return reply;
-  };
+  return mac
+}
 
-  const refresh = () => {
-    set_loaded(false);
-    loadPorts();
-  };
+const getArpReply = (port: number) => {
+  let reply = false
 
-  useEffect(() => {
-    loadPorts();
-  }, []);
+  config.tg_ports.forEach(p => {
+      if(p.port == port) {
+          reply = p.arp_reply ?? false
+      }
+  })
+
+  return reply
+}
+
+const refresh = () => {
+  set_loaded(false)
+  loadPorts()
+}
+
+useEffect(() => {
+  loadPorts()
+
+}, [])
 
   const [currentLanguage, setCurrentLanguage] = useState(
     localStorage.getItem("language") || "en-US"
@@ -187,90 +182,93 @@ const Ports = () => {
           </tr>
         </thead>
         <tbody>
-          {ports.map((v: any, i: number) => {
-            if (loopback_mapping[v["loopback"]] == "Off") {
-              return (
-                <tr key={i}>
-                  <StyledCol className={"col-1"}>{v["pid"]}</StyledCol>
-                  <StyledCol className={"col-1"}>
-                    {v["port"]}/{v["channel"]}
-                  </StyledCol>
-                  <StyledCol className={"col-2"}>{getMac(v["port"])}</StyledCol>
-                  <StyledCol className={"col-2"}>
-                    {speed_mapping[v["speed"] as string]}
-                  </StyledCol>
-                  <StyledCol className={"col-2"}>
-                    <Form.Select
-                      onChange={async (event: any) => {
-                        await updatePort(
-                          v["pid"],
-                          v["speed"],
-                          v["fec"],
-                          event.target.value
-                        );
-                      }}
-                    >
-                      {Object.keys(auto_neg_mapping).map((f) => {
-                        return (
-                          <option selected={f == v["auto_neg"]} value={f}>
-                            {auto_neg_mapping[f]}
-                          </option>
-                        );
-                      })}
-                    </Form.Select>
-                  </StyledCol>
-                  <StyledCol className={"col-2"}>
-                    <Form.Select
-                      onChange={async (event: any) => {
-                        await updatePort(
-                          v["pid"],
-                          v["speed"],
-                          event.target.value,
-                          v["auto_neg"]
-                        );
-                      }}
-                    >
-                      {Object.keys(fec_mapping).map((f) => {
-                        if (
-                          f != "BF_FEC_TYP_FC" ||
-                          v["speed"] != "BF_SPEED_100G"
-                        ) {
-                          return (
-                            <option selected={f == v["fec"]} value={f}>
-                              {fec_mapping[f]}
-                            </option>
-                          );
-                        }
-                      })}
-                    </Form.Select>
-                  </StyledCol>
-                  <StyledCol className={"col-1"}>
-                    <Form.Check
-                      defaultChecked={getArpReply(v["port"])}
-                      onChange={async (event: any) => {
-                        await updateArp(v["pid"], event.target.checked);
-                      }}
-                      type={"switch"}
-                    ></Form.Check>
-                  </StyledCol>
-                  <StyledCol className={"col-1"}>
-                    <PortStatus active={v["status"]} />
-                  </StyledCol>
-                </tr>
-              );
-            }
-          })}
-        </tbody>
-      </Table>
+        {ports.map((v: any, i: number) => {
+                if (loopback_mapping[v["loopback"]] != "On" || p4tg_infos.loopback) {
+                    return <tr key={i}>
+                        <StyledCol className={"col-1"}>{v["pid"]}</StyledCol>
+                        <StyledCol className={"col-1"}>{v['port']}/{v["channel"]}</StyledCol>
+                        <StyledCol className={"col-2"}>{getMac(v['port'])}</StyledCol>
+                        <StyledCol className={"col-2"}>
+                            <Form.Select onChange={async (event: any) => {
+                                let fec = v.fec
 
-      <Row>
-        <Col>
-          <Button onClick={refresh} className={"ml-3"}>
-            <i className="bi bi-arrow-clockwise" />{" "}
-            {translate("buttons.refresh", currentLanguage)}
-          </Button>
-        </Col>
-      </Row>
+                                // 400G requires RS
+                                if(event.target.value == SPEED.BF_SPEED_400G) {
+                                    fec = FEC.BF_FEC_TYP_REED_SOLOMON
+                                }
+
+                                // 10G & 40G does not allow RS
+                                if((event.target.value == SPEED.BF_SPEED_10G || event.target.value == SPEED.BF_SPEED_40G) && v.fec == FEC.BF_FEC_TYP_REED_SOLOMON) {
+                                    fec = FEC.BF_FEC_TYP_NONE
+                                }
+
+                                // 100G does not allow FC
+                                if(event.target.value == SPEED.BF_SPEED_100G && v.fec == FEC.BF_FEC_TYP_FC) {
+                                    fec = FEC.BF_FEC_TYP_NONE
+                                }
+
+                                await updatePort(v.pid, event.target.value, fec, v.auto_neg)
+                            }}>
+                                {Object.keys(speed_mapping).map(f => {
+                                    if(f == SPEED.BF_SPEED_400G && p4tg_infos.asic != ASIC.Tofino2) {
+                                        return
+                                    }
+
+                                    return <option selected={f == v.speed}
+                                                   value={f}>{speed_mapping[f]}</option>
+                                })}
+                            </Form.Select>
+                        </StyledCol>
+                        <StyledCol className={"col-2"}>
+                            <Form.Select onChange={async (event: any) => {
+                                await updatePort(v["pid"], v["speed"], v["fec"], event.target.value)
+                            }}>
+                                {Object.keys(auto_neg_mapping).map(f => {
+                                    return <option selected={f == v["auto_neg"]}
+                                                   value={f}>{auto_neg_mapping[f]}</option>
+                                })}
+                            </Form.Select></StyledCol>
+                        <StyledCol className={"col-2"}><Form.Select onChange={async (event: any) => {
+                            await updatePort(v["pid"], v["speed"], event.target.value, v["auto_neg"])
+                        }}>
+                            {Object.keys(fec_mapping).map(f => {
+                                if(f != FEC.BF_FEC_TYP_REED_SOLOMON && v.speed == SPEED.BF_SPEED_400G) {
+                                    return
+                                }
+
+                                if(f == FEC.BF_FEC_TYP_REED_SOLOMON && (v.speed == SPEED.BF_SPEED_10G ||v.speed == SPEED.BF_SPEED_40G)) {
+                                    return
+                                }
+
+                                if (f != FEC.BF_FEC_TYP_FC || v.speed!= SPEED.BF_SPEED_100G) {
+                                    return <option selected={f == v["fec"]} value={f}>{fec_mapping[f]}</option>
+                                }
+                            })}
+                        </Form.Select>
+                        </StyledCol>
+                        <StyledCol className={"col-1"}>
+                            <Form.Check
+                                defaultChecked={getArpReply(v['port'])}
+                                onChange={async (event: any) => {
+                                    await updateArp(v["pid"], event.target.checked)
+                                }}
+                                type={"switch"}
+                                >
+                            </Form.Check>
+                        </StyledCol>
+                        <StyledCol className={"col-1"}><PortStatus active={v['status']}/></StyledCol>
+                    </tr>
+                }
+            })
+            }
+            </tbody>
+        </Table>
+
+        <Row>
+            <Col>
+                <Button onClick={refresh} className={"ml-3"}><i className="bi bi-arrow-clockwise"/> Refresh</Button>
+            </Col>
+        </Row>
 
       <GitHub />
     </Loader>